+++ src/CancellablePromiseInterface.php
@@ -2,8 +2,19 @@
 
 namespace React\Promise;
 
-interface CancellablePromiseInterface extends PromiseInterface
+/**
+ * This interface is only kept for backward compatibility and must not be used
+ * anymore.
+ *
+ * @deprecated
+ */
+interface CancellablePromiseInterface
 {
+    /**
+     * @return PromiseInterface
+     */
+    public function then(callable $onFulfilled = null, callable $onRejected = null);
+
     /**
      * @return void
      */